add testnet default params

+++ pkgs/consensus/service/bft/config.go
@@ -8,6 +8,6 @@ type BftConfig struct {
 	MyPk           *secp256k1.PublicKey `json:"mypk"`
 	StartMiner     bool                 `json:"startMiner"`
 	ProducerNum    int                  `json:"producerNum"`
-	BlockInterval  int64                `json:"blockInterval"`
+	BlockInterval  int16                `json:"blockInterval"`
 	ChangeInterval uint64               `json:"changeInterval"`
 }
